heroku/requirements.txt
selenium
bs4
deta
python-dotenv
webdriver-manager
apscheduler